Refactor, Cmt: 주석 수정 및 중복로직 삭제
